Reorganize environment directories and add sim visualization
