--- docker/aloha_act/.gitignore
@@ -1,2 +1,0 @@
-ckpts
-data
