--- docker/aloha_act/compose_sim.yml
@@ -1,71 +1,0 @@
-# Run with:
-# docker compose -f docker/aloha_act/compose_sim.yml up --build
-services:
-  aloha_act_sim:
-    image: aloha_act
-    depends_on:
-      - pi0_server
-    build:
-      context: ../..
-      dockerfile: docker/aloha_act/Dockerfile
-    init: true
-    tty: true
-    network_mode: host
-    privileged: true
-    volumes:
-      - /tmp/.X11-unix:/tmp/.X11-unix:ro
-      - ./data:/data
-      - ./ckpts:/ckpts
-    environment:
-      - DISPLAY=$DISPLAY
-    command:
-      - python
-      - imitate_episodes.py
-      - --task_name
-      - sim_transfer_cube_scripted
-      - --ckpt_dir
-      - /ckpts
-      - --eval
-      - --policy_class
-      - PI
-      - --kl_weight
-      - "10"
-      - --chunk_size
-      - "100"
-      - --hidden_dim
-      - "512"
-      - --batch_size
-      - "8"
-      - --dim_feedforward
-      - "3200"
-      - --num_epochs
-      - "2000"
-      - --lr
-      - "1e-5"
-      - --seed
-      - "0"
-      - --render_height
-      - "224"
-      - --render_width
-      - "224"
-      - --onscreen_render
-
-  pi0_server:
-    image: pi0_server
-    build:
-      context: ../..
-      dockerfile: docker/pi0/Dockerfile
-    init: true
-    tty: true
-    network_mode: host
-    volumes:
-      - $PWD:/app
-
-    # Comment out this block if not running on a machine with GPUs.
-    deploy:
-      resources:
-        reservations:
-          devices:
-            - driver: nvidia
-              count: 1
-              capabilities: [gpu]
