+++ examples/aloha_real/real_env.py
@@ -6,8 +6,8 @@ from interbotix_xs_modules.arm import InterbotixManipulatorXS
 from interbotix_xs_msgs.msg import JointSingleCommand
 import numpy as np
 
-from openpi.runtime.environments.aloha_real import constants
-from openpi.runtime.environments.aloha_real import robot_utils
+from examples.aloha_real import constants
+from examples.aloha_real import robot_utils
 
 
 class RealEnv:
