+++ examples/aloha_real/requirements.txt
@@ -1,5 +1,5 @@
 # This file was autogenerated by uv via the following command:
-#    uv pip compile src/openpi/runtime/environments/requirements_ros1.in -o src/openpi/runtime/environments/requirements_ros1.txt --python-version 3.10
+#    uv pip compile src/openpi/runtime/environments/aloha_real/requirements.in -o src/openpi/runtime/environments/aloha_real/requirements.txt --python-version 3.10
 absl-py==2.1.0
     # via
     #   dm-control
