+++ requirements.in
@@ -1,0 +1,19 @@
+augmax
+dm-tree
+einops
+equinox
+flatbuffers
+flax
+imageio
+ipykernel
+jax[cuda12]==0.4.34
+jaxtyping==0.2.34
+numpy==1.26.4
+opencv-python
+orbax-checkpoint
+pillow
+pytest
+sentencepiece
+tqdm
+typing-extensions
+tyro
