+++ requirements.txt
@@ -1,0 +1,256 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements.in -o requirements.txt --python-version 3.11
+absl-py==2.1.0
+    # via
+    #   chex
+    #   optax
+    #   orbax-checkpoint
+asttokens==3.0.0
+    # via stack-data
+augmax==0.3.4
+    # via -r requirements.in
+chex==0.1.87
+    # via optax
+comm==0.2.2
+    # via ipykernel
+debugpy==1.8.9
+    # via ipykernel
+decorator==5.1.1
+    # via ipython
+dm-tree==0.1.8
+    # via -r requirements.in
+docstring-parser==0.16
+    # via tyro
+einops==0.8.0
+    # via
+    #   -r requirements.in
+    #   augmax
+equinox==0.11.10
+    # via -r requirements.in
+etils==1.11.0
+    # via
+    #   optax
+    #   orbax-checkpoint
+executing==2.1.0
+    # via stack-data
+flatbuffers==24.3.25
+    # via -r requirements.in
+flax==0.10.2
+    # via -r requirements.in
+fsspec==2024.10.0
+    # via etils
+humanize==4.11.0
+    # via orbax-checkpoint
+imageio==2.36.1
+    # via -r requirements.in
+importlib-resources==6.4.5
+    # via etils
+iniconfig==2.0.0
+    # via pytest
+ipykernel==6.29.5
+    # via -r requirements.in
+ipython==8.30.0
+    # via ipykernel
+jax==0.4.34
+    # via
+    #   -r requirements.in
+    #   augmax
+    #   chex
+    #   equinox
+    #   flax
+    #   optax
+    #   orbax-checkpoint
+jax-cuda12-pjrt==0.4.34
+    # via jax-cuda12-plugin
+jax-cuda12-plugin==0.4.34
+    # via jax
+jaxlib==0.4.34
+    # via
+    #   chex
+    #   jax
+    #   optax
+jaxtyping==0.2.34
+    # via
+    #   -r requirements.in
+    #   equinox
+jedi==0.19.2
+    # via ipython
+jupyter-client==8.6.3
+    # via ipykernel
+jupyter-core==5.7.2
+    # via
+    #   ipykernel
+    #   jupyter-client
+markdown-it-py==3.0.0
+    # via rich
+matplotlib-inline==0.1.7
+    # via
+    #   ipykernel
+    #   ipython
+mdurl==0.1.2
+    # via markdown-it-py
+ml-dtypes==0.5.0
+    # via
+    #   jax
+    #   jaxlib
+    #   tensorstore
+msgpack==1.1.0
+    # via
+    #   flax
+    #   orbax-checkpoint
+nest-asyncio==1.6.0
+    # via
+    #   ipykernel
+    #   orbax-checkpoint
+numpy==1.26.4
+    # via
+    #   -r requirements.in
+    #   augmax
+    #   chex
+    #   flax
+    #   imageio
+    #   jax
+    #   jaxlib
+    #   ml-dtypes
+    #   opencv-python
+    #   optax
+    #   orbax-checkpoint
+    #   scipy
+    #   tensorstore
+nvidia-cublas-cu12==12.6.4.1
+    # via
+    #   jax-cuda12-plugin
+    #   nvidia-cudnn-cu12
+    #   nvidia-cusolver-cu12
+nvidia-cuda-cupti-cu12==12.6.80
+    # via jax-cuda12-plugin
+nvidia-cuda-nvcc-cu12==12.6.85
+    # via jax-cuda12-plugin
+nvidia-cuda-runtime-cu12==12.6.77
+    # via jax-cuda12-plugin
+nvidia-cudnn-cu12==9.6.0.74
+    # via jax-cuda12-plugin
+nvidia-cufft-cu12==11.3.0.4
+    # via jax-cuda12-plugin
+nvidia-cusolver-cu12==11.7.1.2
+    # via jax-cuda12-plugin
+nvidia-cusparse-cu12==12.5.4.2
+    # via
+    #   jax-cuda12-plugin
+    #   nvidia-cusolver-cu12
+nvidia-nccl-cu12==2.23.4
+    # via jax-cuda12-plugin
+nvidia-nvjitlink-cu12==12.6.85
+    # via
+    #   jax-cuda12-plugin
+    #   nvidia-cufft-cu12
+    #   nvidia-cusolver-cu12
+    #   nvidia-cusparse-cu12
+opencv-python==4.10.0.84
+    # via -r requirements.in
+opt-einsum==3.4.0
+    # via jax
+optax==0.2.4
+    # via flax
+orbax-checkpoint==0.10.2
+    # via
+    #   -r requirements.in
+    #   flax
+packaging==24.2
+    # via
+    #   ipykernel
+    #   pytest
+parso==0.8.4
+    # via jedi
+pexpect==4.9.0
+    # via ipython
+pillow==11.0.0
+    # via
+    #   -r requirements.in
+    #   imageio
+platformdirs==4.3.6
+    # via jupyter-core
+pluggy==1.5.0
+    # via pytest
+prompt-toolkit==3.0.48
+    # via ipython
+protobuf==5.29.1
+    # via orbax-checkpoint
+psutil==6.1.0
+    # via ipykernel
+ptyprocess==0.7.0
+    # via pexpect
+pure-eval==0.2.3
+    # via stack-data
+pygments==2.18.0
+    # via
+    #   ipython
+    #   rich
+pytest==8.3.4
+    # via -r requirements.in
+python-dateutil==2.9.0.post0
+    # via jupyter-client
+pyyaml==6.0.2
+    # via
+    #   flax
+    #   orbax-checkpoint
+pyzmq==26.2.0
+    # via
+    #   ipykernel
+    #   jupyter-client
+rich==13.9.4
+    # via
+    #   flax
+    #   tyro
+scipy==1.14.1
+    # via
+    #   jax
+    #   jaxlib
+sentencepiece==0.2.0
+    # via -r requirements.in
+shtab==1.7.1
+    # via tyro
+simplejson==3.19.3
+    # via orbax-checkpoint
+six==1.17.0
+    # via python-dateutil
+stack-data==0.6.3
+    # via ipython
+tensorstore==0.1.69
+    # via
+    #   flax
+    #   orbax-checkpoint
+toolz==1.0.0
+    # via chex
+tornado==6.4.2
+    # via
+    #   ipykernel
+    #   jupyter-client
+tqdm==4.67.1
+    # via -r requirements.in
+traitlets==5.14.3
+    # via
+    #   comm
+    #   ipykernel
+    #   ipython
+    #   jupyter-client
+    #   jupyter-core
+    #   matplotlib-inline
+typeguard==2.13.3
+    # via jaxtyping
+typing-extensions==4.12.2
+    # via
+    #   -r requirements.in
+    #   chex
+    #   equinox
+    #   etils
+    #   flax
+    #   ipython
+    #   orbax-checkpoint
+    #   tyro
+tyro==0.9.1
+    # via -r requirements.in
+wcwidth==0.2.13
+    # via prompt-toolkit
+zipp==3.21.0
+    # via etils
