--- src/openpi/runtime/environments/aloha_sim/README.md
@@ -1,13 +1,0 @@
-# Run Aloha Sim
-
-Termina window 1:
-
-```
-uv run src/openpi/runtime/environments/aloha_sim/main.py
-```
-
-Termina window 2:
-
-```
-uv run src/openpi/serving/http_policy_server_pi0.py --mode SIM
-```
